// src/types.ts
export enum StudioView {
    Book = 'book',
    Chat = 'chat',
    Image = 'image',
    Video = 'video',
    Audio = 'audio',
    Library = 'library',
    AIPresenter = 'ai_presenter',
    AIInfluencer = 'ai_influencer',
    AIAvatar = 'ai_avatar',
    AIPhotoStudio = 'ai_photo_studio',
    AIHyperRealisticPhotos = 'ai_hyper_realistic_photos',
    AIConversationalAvatar = 'ai_conversational_avatar',
    AIHeadshotGenerator = 'ai_headshot_generator',
    AIVirtualBackground = 'ai_virtual_background',
    AIDesigner = 'ai_designer',
    AIMusicGenerator = 'ai_music_generator',
    AISoundEffectsGenerator = 'ai_sound_effects_generator',
    MagicUpscaler = 'magic_upscaler',
    BrandVoice = 'brand_voice',
    BlogPostWorkflow = 'blog_post_workflow',
    LongFormEditor = 'long_form_editor',
    InstagramCaption = 'instagram_caption',
    AITextHumanizer = 'ai_text_humanizer',
    SEOTitleGenerator = 'seo_title_generator',
    MetaDescriptionGenerator = 'meta_description_generator',
    KeywordGenerator = 'keyword_generator',
    FacebookPostGenerator = 'facebook_post_generator',
    ContentIdeaGenerator = 'content_idea_generator',
    YoutubeToArticle = 'youtube_to_article',
    YoutubeTitleGenerator = 'youtube_title_generator',
    YoutubeDescriptionGenerator = 'youtube_description_generator',
    YoutubeIdeaGenerator = 'youtube_idea_generator',
    YoutubeScriptOutlineGenerator = 'youtube_script_outline_generator',
    AboutUsPageGenerator = 'about_us_page_generator',
    CompanyProfileGenerator = 'company_profile_generator',
    CompanyBioGenerator = 'company_bio_generator',
    AboutMeGenerator = 'about_me_generator',
    GrammarChecker = 'grammar_checker',
    PolicyGenerator = 'policy_generator',
    MarketingCampaignGenerator = 'marketing_campaign_generator',
    ExplainLikeImFive = 'explain_like_im_five',
    SeriesStoryGenerator = 'series_story_generator',
    TikTokCaptionGenerator = 'tiktok_caption_generator',
    TikTokHashtagGenerator = 'tiktok_hashtag_generator',
    GoogleAdsTitleGenerator = 'google_ads_title_generator',
    GoogleAdsDescriptionGenerator = 'google_ads_description_generator',
    GoogleMyBusinessProductGenerator = 'google_my_business_product_generator',
    GoogleMyBusinessPostGenerator = 'google_my_business_post_generator',
    FacebookAdsTitleGenerator = 'facebook_ads_title_generator',
    FacebookAdsPrimaryTextGenerator = 'facebook_ads_primary_text_generator',
    SocialMediaPostGenerator = 'social_media_post_generator',
    MusicIdeaGenerator = 'music_idea_generator',
    SongLyricsGenerator = 'song_lyrics_generator',
    BookTitleGenerator = 'book_title_generator',
    BookOutlineGenerator = 'book_outline_generator',
    ShakespeareGenerator = 'shakespeare_generator',
    AIStoryGenerator = 'ai_story_generator',
    RomanceStoryGenerator = 'romance_story_generator',
    StoryPlotGenerator = 'story_plot_generator',
    FanficCreator = 'fanfic_creator',
    ManuscriptWriter = 'manuscript_writer',
    BusinessPlanGenerator = 'business_plan_generator',
    CharacterDescriptionGenerator = 'character_description_generator',
    CharacterBackstoryGenerator = 'character_backstory_generator',
    TextToneAnalyzer = 'text_tone_analyzer',
    Settings = 'settings',
    FAQ = 'faq',
    Disclaimer = 'disclaimer',
    UsagePolicy = 'usage_policy',
    Plans = 'plans',
}